+++ R/generateNCI.R
@@ -31,7 +31,7 @@ NULL
 #' @examples
 #' generateNCI()
 generateNCI <- function(
-  grid = 50, # grid size
+  grid = 20, # grid size
   Nt = 50, # number of time-steps
   timestep = 30, # time-step length in years
   muNCI = 124, # mu of normal distribution for NCI
@@ -43,7 +43,7 @@ generateNCI <- function(
   lambda = 0.31, # lambda of exponential distribution for negative deltaNCI
   d = 3 # spatial auto-correlation size (3*3m)
 ){
-  Var1 <- Var2 <- value <- NULL
+  Var1 <- Var2 <- value <- Year <- NULL
   
   nci_y <- rnorm(grid*grid, 
                  mean = muNCI, 
@@ -65,7 +65,7 @@ generateNCI <- function(
                                                                mu = mu, 
                                                                sigma = sigma, 
                                                                lambda = lambda)
-    if(y %in% (timestep*(1:50))){
+    if(y %in% (timestep*(1:Nt))){
       nci <- bind_rows(nci,  
                        as.matrix(disaggregate(nci_y, d, method = "bilinear"))[1:grid,1:grid] %>%
                          melt() %>%
@@ -75,69 +75,11 @@ generateNCI <- function(
     }
   }
     
-  return(nci)
+  return(nci %>% 
+           group_by(Year) %>% 
+           mutate(Ind = 1:n()) %>% 
+           ungroup %>% 
+           dcast(Year ~ Ind, value.var = "NCI") %>% 
+           select(-Year) %>% 
+           as.matrix())
 }
-
-#' Wrapper for generateNCI
-#' 
-#' Format in a matrix for the simulator.
-#'
-#' @param grid int. Number of cells per side of the matrix.
-#' @param Nt int. Number of time-steps.
-#' @param timestep int. Time-step length in years.
-#' @param muNCI double. mu parameter for the normal distribution used for NCI.
-#' @param sigmaNCI double. sigma parameter for the normal distribution used for
-#'   NCI.
-#' @param alpha double. Intercept for the Bernoulli distribution determining the
-#'   risk to have a negative deltaNCI.
-#' @param beta double. Slope of previous NCI for the Bernoulli distribution
-#'   determining the risk to have a negative deltaNCI.
-#' @param mu double. mu parameter for the lognormal distribution used for
-#'   positive deltaNCI.
-#' @param sigma double. sigma parameter for the lognormal distribution used for
-#'   positive deltaNCI.
-#' @param lambda double. lambda parameter for the exponential distribution used
-#'   for negative deltaNCI.
-#' @param d int. Spatial autoccorelation size in number of cells.
-#'
-#' @return A matrix..
-#' 
-#' @export
-#'
-#' @examples
-#' generateNCIsim()
-generateNCIsim <- function(
-  grid = 50, # grid size
-  Nt = 50, # number of timesteps
-  timestep = 30, # time-step length in years
-  muNCI = 0.749, # mu of normal distributon for nci
-  sigmaNCI = 2.651, # sigma of normal distributon nci
-  alpha = -1.32,
-  beta = 0.003,
-  mu = 0.749, # mu of lognormal distributon for positive deltanci
-  sigma = 2.651, # sigma of lognormal distributon for positive deltanci
-  lambda = 0.31, # lamnda of exponential distributon for negative deltanci
-  d = 3 # spatial autocorrelation size (3*3m)
-){
-  Year <- NULL
-  nci <- generateNCI(
-    grid = grid, 
-    Nt = Nt, 
-    timestep = timestep,
-    muNCI = muNCI, 
-    sigmaNCI = sigmaNCI, 
-    alpha = alpha,
-    beta = beta,
-    mu = mu,
-    sigma = sigma,
-    lambda = lambda,
-    d = d)
-  nci %>% 
-    group_by(Year) %>% 
-    mutate(Ind = 1:n()) %>% 
-    ungroup %>% 
-    dcast(Year ~ Ind, value.var = "NCI") %>% 
-    select(-Year) %>% 
-    as.matrix()
-}
-