+++ R/simulator.R
@@ -1,35 +1,16 @@
-#' @include utils-pipe.R generateNCI.R RcppExports.R sinusoidalTopography.R
-#'   squareDiamondTopography.R paracouTopography.R
+#' @include utils-pipe.R generateNCI.R RcppExports.R generateTopography.R
 #' @importFrom dplyr bind_rows left_join rename
 #' @importFrom reshape2 melt
 #' @importFrom tidyr separate
 NULL
 
-#' simulator
+#' Simulator
 #'
 #' @param grid int.  Number of cells per side of the matrix.
 #' @param Nt int. Number of time steps.
 #' @param timestep int. Time-step length in years.
-#' @param topography char. Topography generator between 'sinusoidal',
-#'   'squarediamond' or 'paracou'.
-#' @param Elim double. Environmental matrix extrme (absolute value).
-#' @param amplitude double. Amplitude of the sinusoidal functional.
-#' @param ruggedness double. Ruggedness parameter.
-#' @param muNCI double. mu parameter for the normal distribution used for NCI.
-#' @param sigmaNCI double. sigma parameter for the normal distribution used for
-#'   NCI.
-#' @param alpha double. Intercept for the Bernoulli distribution determining the
-#'   risk to have a negative deltaNCI.
-#' @param beta double. Slope of previous NCI for the Bernoulli distribution
-#'   determining the risk to have a negative deltaNCI.
-#' @param mu double. mu parameter for the lognormal distribution used for
-#'   positive deltaNCI.
-#' @param sigma double. sigma parameter for the lognormal distribution used for
-#'   positive deltaNCI.
-#' @param lambda double. lambda parameter for the exponential distribution used
-#'   for negative deltaNCI.
-#' @param d int. Spatial auto-correlation size in number of cells.
-#' @param plot int. Plot number between 1 and 15 (a cell size is 3x3m).
+#' @param topography matrix. Topography matrix generated with generateTopography.
+#' @param NCI matrix. Topography matrix generated with generateNCIsim.
 #' @param sigmaGtopo double. Variance of genetic values with topography.
 #' @param sigmaZtopo double. Plasticity of phenotypes with topography.
 #' @param sigmaGnci double. Variance of genetic values with NCI.
@@ -39,7 +20,6 @@ NULL
 #' @param Rpollination int. Pollination radius in cells (father to mother).
 #' @param Rdispersion int. Dispersal radius in cells (mother to seedling).
 #' @param determinist bool. Deterministic or probabilistic viability.
-#' @param verbose bool. Should the function print statuses.
 #'
 #' @return A data frame.
 #'
@@ -53,19 +33,8 @@ simulator <- function(
   grid = 20,
   Nt = 50,
   timestep = 30, # time-step length in years
-  topography = "sinusoidal",
-  Elim = 5,
-  amplitude = 1,
-  ruggedness = 1,
-  plot = 1,
-  muNCI = 124, # mu of normal distribution for NCI
-  sigmaNCI = 26, # sigma of normal distribution NCI
-  alpha = -1.32,
-  beta = 0.003,
-  mu = 0.749, # mu of lognormal distribution for positive deltaNCI
-  sigma = 2.651, # sigma of lognormal distribution for positive deltaNCI
-  lambda = 0.31, # lambda of exponential distribution for negative deltaNCI
-  d = 3, # spatial auto-correlation size (3*3m)
+  topography = generateTopography(),
+  NCI = generateNCI(),
   sigmaGtopo = 1,
   sigmaZtopo = 1,
   sigmaGnci = 26,
@@ -74,41 +43,17 @@ simulator <- function(
   Ns = 4,
   Rpollination = 1,
   Rdispersion = 1,
-  determinist = TRUE,
-  verbose = TRUE
+  determinist = TRUE
 ){
   Var1 <- Var2 <- var <- NULL
-  
-  # Topo
-  if(verbose) message("Generating topography.")
-  if(!(topography %in% c('sinusoidal', 'squarediamond', 'paracou')))
-    stop(paste(topography, "option doesn't exist to generate topography (see help)."))
-  if(topography == 'sinusoidal')
-    Topo <- sinusoidalTopography(grid = grid, Elim = Elim, amplitude = amplitude)
-  if(topography == 'squarediamond')
-    Topo <- squareDiamondTopography(grid = grid,  ruggedness = ruggedness)
-  if(topography == 'paracou')
-    Topo <- paracouTopography(grid = grid, plot = plot, Elim = Elim)
-  
-  # NCI
-  if(verbose) message("Generating NCI.")
-  NCI <- generateNCIsim(
-    grid = grid, 
-    Nt = Nt, 
-    timestep = timestep,
-    muNCI = muNCI, 
-    sigmaNCI = sigmaNCI, 
-    alpha = alpha,
-    beta = beta,
-    mu = mu,
-    sigma = sigma,
-    lambda = lambda,
-    d = d)
-  
-  # Sim
-  if(verbose) message("Running simulation.")
+  if(!all(dim(topography) == grid))
+    stop(paste("Topography is not of size grid =", grid, "but of size", dim(topography)[1], "x", dim(topography)[2]))
+  if(ncol(NCI) != grid*grid)
+    stop(paste("Columns of NCI are not of size grid*grid =", grid*grid, "but of size", ncol(NCI)))
+  if(nrow(NCI) != Nt)
+    stop(paste("Rows of NCI are not of size Nt =", Nt, "but of size", nrow(NCI)))
   sim <- simulatorCpp(
-    Topo = Topo,
+    Topo = topography,
     NCI = NCI,
     grid = grid, 
     Nt = Nt,